Some reworking of things...
Yet another redesign of the bonus enemey spawn system. This will likely
be it's final form. :P

Fixed bugs with PlaceWallMimics() not placing Wall Mimics on south
walls. This was a bug with original code by DodgeRoll. :P

Another bug was PlaceWallMimics placing wall mimics on thin walls
dividing 2 rooms. This would cause issues and WallMimicController gets
confused and tries to spawn the mimic in teh wrong room. It pops out the
wall like normal but then immediately vanishes and no longer appears in
either room. I've limited it not place mimics in walls that have floor
tiles on the other side. This should take care of any instance of this
happening.

Also maybe fixed other things. But forgot the details.