Reworked Object Instantiation, fixed Challenge Mode.
* Restored random challenge mode rooms. Found away to disable challange
mod when a teleport ouf the room occurs. Softlocks due to players being
unable to complete a room due to certain modifiers from a previous room
should no longer be an issue.

* Reworked how objects are instanced in Object Randomizer. Forge Hammers
now work properly and are now functional when placed in random rooms.
"Kickable" Tables that are spawned are no longer causing existing tables
on future floors being changed into kickable objects.

* A few more NPCs added to NPC spawn pool.

* Fixed issue with "phantom" loot drop icon on minimap on most floors.
This was caused by the folding table object and how I instanced that
one. I don't know where the GameObject asset is for that one so for now
that one has been removed from the spawn pool.

* Lunk removed from NPC spawn pool. He doesn't really work properly and
was generating exceptions in the logs. The extra NPCs I added make up
for his removable. :P