Preperation for 2.0 release...
* Reworked Object Randomizer Placement code. It now caches previously
placed objects. This should result in less objects overlapping another
object.
* Added Heart Dispenser to Object Randomizer. They even work if they
happen to spawn on the Forge. ;)
* Disabled collision for certain objects placed by Object Randomizer.
Possiblity of NPCs/Locked Doors blocking room entrance/exit ways is no
longer a problem.
* Added new "Hologram" shader effect to random enemies in Ultra mode.
Enemies that become "holograms" can no longer collide with the player.
* Disabled primary collision collider of critters like snakes, chickens,
rats, robot past critters so that enemies no longer get stuck on them.
As a consequence rats and chickens must be shot with a gun to be killed
although player no longer needs to worry about them being jammed as they
can no longer contact the player.
* In Ultra mode, random objects in rooms can become holograms. Excluding
certain objects like doors, this results in them no longer colliding
with the player.
* Player is granted iframes during teleport in tentacle time to prevent
damage in certain circumstances like if being teleported while a Forge
Hammer is coming down or while on poison/fire goop.
* Tentacle Teleport now properly stuns enemies in the destination room
if teleporting to a different room instead of only stunning enemeis in
the room the player is leaving.
* Added console command for disabling explosion nerf introduced by AG&D.
Use "chaos fixexplosions" to enable/disable.
* Added console command to allow calling Tentacle time on demand.
Specifically the version that teleports the player to another room.
This may be useful to escape a room in the event an unforseen bug causes
the room to not end combat properly.