Update for FTA support....
* All the features of the mod should work in the new Farewell to Arms
update.
* First full custom DungeonFlow created! It is a recreation of
TEST_West_Floor_03a_Flow a flow that was removed in the AG&D update. I
have brought it back and modified it for use as the glitch chest floor.
Opening glitch chests will now take you to this floor where there will
be glitched objects and enemies. :D
* Secret floor modified to use the new tileset. Also fixed to work on
FTA update. Dodge Roll moved all the room prefabs for Catacombs
(internal name for the 4th floor) into the dungeon asset bundle for that
floor. So I had to change where I was loading those assets.
* Added soem code to hopefully mitigate a lag issue the new glitched
trap door object may cause if player opens the glitched trap door for
the secret floor early during a run on the floor the object appears on.
* For balance reasons, All extra wall mimic code has been disabled due
to FTA update adding item drops to Wall mimics. I haven't figured out
how to prevent this yet, so for now the mod will use the standard spawn
odds of the base game and will not increase occurance of wall mimics in
any of the chaos modes. Commands relating to the Wall mimics have also
been disabled.
* Some of the new enemies introduced for the new floor have been added
to Sewers and Abbey secret floors as the new designs would make them a
fitting addition to those floors.
* Visiting the secret floor will cause a bunch of the new enemies to
appear on the Forge and Bullet Hell. Note this is setup to only occur
once in a run so once you start a new run, it won't happen again unless
you decide to visit the secret floor again.