Some Bug Fixes...
* Fixed randomization of boss room and winchester room on custom secret
floor.
* Prevented bosses/minibosses from getting randomized shader effects in
chaos modes. This caused an exception with the introdoer for Cannon
Balrag that softlocked the game. I'm not sure why this happens but it
should not occur anymore (hopefully. The bug was intermittet so I can't
test this reliably)
* Fixed missing console text message for "chaos glitch" command.
* Fixed incorrect text for "chaos walls" command. It now correctly
indicates the correct state of the wall mimic mode being deactivated or
activated.
* Disabled Raccoon's custom behavior. This is available via another mod
now and I've disabled this to prevent conflicts.